delete_order.py handler is singled out.
